Introduce View type and Max/Min overloading
View is a new type that represents a static view of a population. View
is simply a named type for []Genome.

Population.Members() is now named Population.View() and returns a View.

The functions Max(...Genome) and Min(...Genome) have been replaced in
favor of methods on the Population interface. A new function,
Tournament(...Genome), was (re)introduced to take over the old use
cases of Max(...Genome).