Settle on a good Close() policy
Genomes may control resources that need to be closed. This is
especially true when using populations as genomes in hybrid techniques.

The Close() policy is this: Whenever a population "kills" a genome, it
calls the Close() method of that genome. This can occur from two
different cases. Firstly, when Close() is called on a population, it
calls Close() on all of its members. Secondly, whenever a genome is
replaced within the population, it calls Close() on the genome being
replaced.

This effectively reverts 1e4cd15e1f72af9e55942d3846ba79b029f901a0