The View type is reworked. It becomes a struct type.
Population statistics are now the responsibility of the View type.
The reasoning for this is that (a) statistics make more sense in the
context of a static population and (b) all the statistics can be
computed in linear time when the view is created.

Views now properly handle hybridization. Upon construction of a View,
we descend the population tree until we reach the leaf genomes.

Creating a view requires at least one array allocation. Since
statistics are often used in the user's main loop (e.g. for termination
conditions) we want to minimize that allocation cost. Future work may
be done to pool and reuse Views in an effort to minimize this cost.