Pool views for reuse
Creating a view from scratch requires at least one slice allocation to
store the members. To minimize the allocation cost when repeatedly
creating views, we allow old views to be reused. A reallocation still
occurs when the slice is too small, but eventually all views in the
pool will grow to the maximum capacity and require no reallocation.

In a non-hybrid setup, there will likely be only one view in use at a
time, always with the same capacity. Thus we need only one allocation
until the GC decides to empty the pool.