Close method should not return an error
We have no need to return errors from the close method. None of the
built-in genomes/populations need to return errors, and handling errors
is a lot of unnecessary work. What would it mean if an error occurred?
How would the user recover? Most likely, errors are truly exceptional,
and the algorithm would need to stop. Thus panics are more appropriate.