Code sprint: n-queens
This commit reflects the outcome of a code sprint while writing a
solver for the n-queens problem. This is all new feature code that
should be reviewed (later).

I've introduced Order crossover, Cycle crossover, and PMX as helper
functions for dealing with permutations. BinaryTournament has been
tweaked as well.

The "diffusion population" now has the more general name "graph
population". A delay may be set between iterations, allowing a graph
population to implement an island model when used as a meta-population.

A new n-queens example exists that demonstrates these features. It
implements an island model where each island is a hypercube diffusion
population. The islands are organized in a ring and communicate once
per second. A permutation representation is used with cycle crossover.
A solution to the 256-queens problem was found in a couple minutes
using 4 threads on an Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz.

Everything could use better documentation. The Ackley example is
probably busted, but simple renaming should be enough to fix it.