Major rewrite
Issue #1 revealed some important design limitations. This commit
replaces the view interface with separate iterator and statistics
interfaces. This makes writing populations easier.

The close method is now handled as a finalizer. This means that the
close method is not guaranteed to be called, but it handles the major
use cases. See https://godoc.org/runtime#SetFinalizer

The populations were rewritten for the new interface. The migration
strategy for the graph population (i.e. the Cross method) was changed
to uniform random. I plan to use that same strategy for all
populations. Currently, the generational population panics on cross.

The permutation crossover operation no longer allocate a child gene.
Instead they take a slice allocated by the caller. This allows for
better gene recycling.