Drain updates queue before closing
The updates queue was not being drained in the generational population.
This lead to situations where, upon closing the population, leftover
mating routines would push into the queue, but the channel would be
closed. We make sure to drain the queue so that we are sure all
goroutines are dead.