Run examples with `go test`
The "examples" feature of the testing package was not used, opting to provide
the main functions as "tests" instead. This decision was made for the following
reasons:

- I want to use the example code as unit test. Examples must test against and
exact value on stdout, but Evolutionary algorithms are inherently stochastic.
Meaningful values cannot be tested exactly, while unit tests allow for covering
a greater class of logic errors.

- Examples are unfit for interactive use because they MUST test a specific
output on stdout. Thus the example code cannot print information describe a
stochastic process. On the other hand, full tests allow us to use stdout
directly. We can even print a live status bar during the run.

There is a benefit of examples over full tests in that examples can be displayed
in godoc output, making them ideal for exposure. This can be resolved by
directing the user to the examples package through other means.