Improve tsp example
The example now uses the ATT48 dataset from TSPLIB. The example highlights
hybridization with local search by using the 2-opt heuristic search as a
mutation. We get within 10% of the true minimum after about 300,000 fitness
evaluations.

The search is slower than a simple 2-opt hillclimber, but the goal is to
demonstrate local search hybrids, not to outperform hillclimbing.