Update the core API
The Genome interface is much simpler. The Close hook has been removed and the
Cross method has been replaced by the EvolveFn type. This separates the notion
of the evolve loop from the representation of the genomes. Now Genomes only have
one method, Fitness.

Iterators have been removed. The genomes are evolved in-place, so looping over
the slice is equivalent to an iterator. It's not safe to iterate while the
evolution is running. I'm still thinking about the best way to handle those
use-cases.