Update the populations to match the new API
This change is very substantial. While porting to the new API, I took the time
to review and update most of the code in the pop package.

A new function, Migrate, is introduced in the pop/gen package to help build
migration functions when constructing island models. The delays are handled by
the migration functions, and the SetDelay methods are removed.